Move deprecated vars
