+++ data/colors_v2/utils/deprecated_vars.ts
@@ -1,5 +1,6 @@
-import {get} from '../../src/utils'
+import {get} from '../../../src/utils'
 
+// Variables to be removed in the next major release
 export default {
   text: {
     primary: get('fg.default'),
